+++ login/src/Login.js
@@ -1,0 +1,62 @@
+// Login.js
+import { useState } from "react";
+import "./App.css";
+
+const Login = () => {
+    const initialValues = {
+        email: "",
+        password: "",
+    };
+    const [formValues, setFormValues] = useState(initialValues);
+    const [formErrors, setFormErrors] = useState({});
+    const [isSubmit, setIsSubmit] = useState(false);
+
+    const handleChange = (e) => {
+        const { name, value } = e.target;
+        setFormValues({ ...formValues, [name]: value });
+    };
+
+    const handleSubmit = (e) => {
+        e.preventDefault();
+        // Add validation logic if needed
+        console.log(formValues);
+        setIsSubmit(true);
+    };
+
+    return (
+        <div className="container">
+            <form onSubmit={handleSubmit}>
+                <h1>Login</h1>
+                <div className="ui divider"></div>
+                <div className="ui form">
+                    <div className="field">
+                        <label>Email</label>
+                        <input
+                            type="text"
+                            name="email"
+                            placeholder="Email"
+                            value={formValues.email}
+                            onChange={handleChange}
+                        />
+                    </div>
+                    <div className="field">
+                        <label>Password</label>
+                        <input
+                            type="password"
+                            name="password"
+                            placeholder="Password"
+                            value={formValues.password}
+                            onChange={handleChange}
+                        />
+                    </div>
+                    <button className="fluid ui button blue">Login</button>
+                </div>
+            </form>
+            <div className="text">
+                Don't have an account? <span><a href="/signup">Sign Up</a></span>
+            </div>
+        </div>
+    );
+};
+
+export default Login;
