+++ login/src/SignUp.js
@@ -1,0 +1,120 @@
+// SignUp.js
+import { useState, useEffect } from "react";
+import "./App.css";
+
+const SignUp = () => {
+    const initialValues = {
+        username: "",
+        email: "",
+        password: "",
+        confirmPassword: "",
+    };
+    const [formValues, setFormValues] = useState(initialValues);
+    const [formErrors, setFormErrors] = useState({});
+    const [isSubmit, setIsSubmit] = useState(false);
+
+    const handleChange = (e) => {
+        const { name, value } = e.target;
+        setFormValues({ ...formValues, [name]: value });
+    };
+
+    const handleSubmit = (e) => {
+        e.preventDefault();
+        setFormErrors(validate(formValues));
+        setIsSubmit(true);
+    };
+
+    useEffect(() => {
+        if (Object.keys(formErrors).length === 0 && isSubmit) {
+            console.log(formValues);
+        }
+    }, [formErrors, formValues, isSubmit]);
+
+    const validate = (values) => {
+        const errors = {};
+        const regex = /^[^\s@]+@[^\s@]+\.[^\s@]{2,}$/i;
+        if (!values.username) {
+            errors.username = "Username is required!";
+        }
+        if (!values.email) {
+            errors.email = "Email is required!";
+        } else if (!regex.test(values.email)) {
+            errors.email = "This is not a valid email format!";
+        }
+        if (!values.password) {
+            errors.password = "Password is required";
+        } else if (values.password.length < 4) {
+            errors.password = "Password must be more than 4 characters";
+        } else if (values.password.length > 10) {
+            errors.password = "Password cannot exceed more than 10 characters";
+        }
+        if (values.password !== values.confirmPassword) {
+            errors.confirmPassword = "Those passwords didn’t match. Try again.";
+        }
+        return errors;
+    };
+
+    return (
+        <div className="container">
+            {Object.keys(formErrors).length === 0 && isSubmit ? (
+                <div className="ui message success">Signed up successfully</div>
+            ) : null}
+            <form onSubmit={handleSubmit}>
+                <h1>Sign Up</h1>
+                <div className="ui divider"></div>
+                <div className="ui form">
+                    <div className="field">
+                        <label>Username</label>
+                        <input
+                            type="text"
+                            name="username"
+                            placeholder="Choose a username"
+                            value={formValues.username}
+                            onChange={handleChange}
+                        />
+                    </div>
+                    <p>{formErrors.username}</p>
+                    <div className="field">
+                        <label>Email</label>
+                        <input
+                            type="text"
+                            name="email"
+                            placeholder="Email"
+                            value={formValues.email}
+                            onChange={handleChange}
+                        />
+                    </div>
+                    <p>{formErrors.email}</p>
+                    <div className="field">
+                        <label>Password</label>
+                        <input
+                            type="password"
+                            name="password"
+                            placeholder="Password"
+                            value={formValues.password}
+                            onChange={handleChange}
+                        />
+                    </div>
+                    <p>{formErrors.password}</p>
+                    <div className="field">
+                        <label>Confirm Password</label>
+                        <input
+                            type="password"
+                            name="confirmPassword"
+                            placeholder="Confirm password"
+                            value={formValues.confirmPassword}
+                            onChange={handleChange}
+                        />
+                    </div>
+                    <p>{formErrors.confirmPassword}</p>
+                    <button className="fluid ui button blue">Submit</button>
+                </div>
+            </form>
+            <div className="text">
+                Already have an account? <span><a href="/login">Login</a></span>
+            </div>
+        </div>
+    );
+};
+
+export default SignUp;
